6. DTO [#504800 #379509]: Убрать из контроллера конструкции try-catch

+++ src/main/java/ru/job4j/controller/PersonController.java
@@ -4,22 +4,15 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.domain.Person;
 import ru.job4j.dto.PersonDto;
-import ru.job4j.exception.InvalidPasswordFormatException;
 import ru.job4j.handler.GlobalExceptionHandler;
 import ru.job4j.service.PersonService;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.HashMap;
 import java.util.List;
 
 @RestController
@@ -58,36 +51,34 @@ public class PersonController {
 
     @PostMapping("/")
     public ResponseEntity<Person> create(@RequestBody Person person) {
-        try {
-            return new ResponseEntity<>(
-                    personService.save(person),
-                    HttpStatus.CREATED
-            );
-        } catch (DataIntegrityViolationException e) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
+        return new ResponseEntity<>(
+                personService.save(person),
+                HttpStatus.CREATED
+        );
     }
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Person person) {
-        try {
-            personService.save(person);
-            return ResponseEntity.ok().build();
-        } catch (DataIntegrityViolationException e) {
-            return ResponseEntity.badRequest().build();
+        if (!personService.existsById(person.getId())) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND,
+                    String.format("Person not found for the following ID:  %d", person.getId())
+            );
         }
+        personService.save(person);
+        return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
-        try {
-            personService.deleteById(id);
-            return ResponseEntity.ok().build();
-        } catch (EmptyResultDataAccessException e) {
-            return ResponseEntity.notFound().build();
-        } catch (DataIntegrityViolationException e) {
-            return ResponseEntity.badRequest().build();
+        if (!personService.existsById(id)) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND,
+                    String.format("Person not found for the following ID:  %d", id)
+            );
         }
+        personService.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 
     @PatchMapping("/{id}/change_login")
@@ -100,12 +91,8 @@ public class PersonController {
                         )
                 );
         person.setLogin(personDto.getLogin());
-        try {
-            personService.save(person);
-            return ResponseEntity.ok().build();
-        } catch (DataIntegrityViolationException e) {
-            return ResponseEntity.badRequest().build();
-        }
+        personService.save(person);
+        return ResponseEntity.ok().build();
     }
 
     @PatchMapping("/{id}/change_password")
@@ -118,22 +105,7 @@ public class PersonController {
                         )
                 );
         person.setPassword(personDto.getPassword());
-        try {
-            personService.save(person);
-            return ResponseEntity.ok().build();
-        } catch (DataIntegrityViolationException e) {
-            return ResponseEntity.badRequest().build();
-        }
-    }
-
-    @ExceptionHandler(value = {InvalidPasswordFormatException.class})
-    public void exceptionHandler(Exception e, HttpServletRequest request, HttpServletResponse response) throws IOException {
-        response.setStatus(HttpStatus.BAD_REQUEST.value());
-        response.setContentType("application/json");
-        response.getWriter().write(objectMapper.writeValueAsString(new HashMap<>() { {
-            put("message", e.getMessage());
-            put("type", e.getClass());
-        }}));
-        LOGGER.error(e.getLocalizedMessage());
+        personService.save(person);
+        return ResponseEntity.ok().build();
     }
 }
